backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
python-dotenv==1.0.0
google-generativeai==0.8.0
pymupdf==1.23.8
pillow==10.2.0
httpx==0.26.0
